# backend/api/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: Database API for Nefryt LDS
  title: Nefryt LDS API
  version: 1.0.0
servers:
- description: Test server
  url: http://192.168.1.231:8080/
- description: Developer's local server
  url: http://localhost:8080/
security:
- bearerAuth: []
paths:
...
  /link:
    get:
      description: List all links
      operationId: list_links
      parameters:
        - description: Page number
          in: query
          name: $page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - description: Page size
          in: query
          name: $size
          required: false
          schema:
            type: integer
            maximum: 10000
            minimum: 1
            default: 50
        - description: Query filter in OData standard
          explode: true
          in: query
          name: $filter
          required: false
          schema:
            type: string
          style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Link'
                type: array
          description: A array of links
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List links
      tags:
      - links
      x-openapi-router-controller: api.controllers.links_controller
    post:
      description: Create a links
      operationId: create_link
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Link'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Create links
      tags:
      - links
      x-openapi-router-controller: api.controllers.links_controller
  /link/{linkId}:
    delete:
      description: Deletes specific link
      operationId: delete_link_by_id
      parameters:
        - description: The id of the link to retrieve
          explode: false
          in: path
          name: linkId
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "204":
          description: No content
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Deletes link
      tags:
      - links
      x-openapi-router-controller: api.controllers.links_controller
    get:
      description: Info for specific link
      operationId: get_link_by_id
      parameters:
      - description: The id of the link to retrieve
        explode: false
        in: path
        name: linkId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Gets link details
      tags:
      - links
      x-openapi-router-controller: api.controllers.links_controller
    put:
      description: Updates a links
      operationId: update_link
      parameters:
        - description: The id of the link to retrieve
          explode: false
          in: path
          name: linkId
          required: true
          schema:
            type: integer
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Link'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Updates links
      tags:
      - links
      x-openapi-router-controller: api.controllers.links_controller
  /node:
    get:
      description: List all nodes
      operationId: list_nodes
      parameters:
        - description: Page number
          in: query
          name: $page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - description: Page size
          in: query
          name: $size
          required: false
          schema:
            type: integer
            maximum: 10000
            minimum: 1
            default: 50
        - description: Query filter in OData standard
          explode: true
          in: query
          name: $filter
          required: false
          schema:
            type: string
          style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Node'
                type: array
          description: A array of nodes
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List nodes
      tags:
      - nodes
      x-openapi-router-controller: api.controllers.nodes_controller
    post:
      description: Create a nodes
      operationId: create_node
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Node'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Create nodes
      tags:
      - nodes
      x-openapi-router-controller: api.controllers.nodes_controller
  /node/{nodeId}:
    delete:
      description: Deletes specific node
      operationId: delete_node_by_id
      parameters:
      - description: The id of the node to retrieve
        explode: false
        in: path
        name: nodeId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: No content
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Deletes node
      tags:
      - nodes
      x-openapi-router-controller: api.controllers.nodes_controller
    get:
      description: Info for specific node
      operationId: get_node_by_id
      parameters:
      - description: The id of the node to retrieve
        explode: false
        in: path
        name: nodeId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Gets node details
      tags:
      - nodes
      x-openapi-router-controller: api.controllers.nodes_controller
    put:
      description: Updates a nodes
      operationId: update_node
      parameters:
      - description: The id of the node to retrieve
        explode: false
        in: path
        name: nodeId
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Node'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Updates nodes
      tags:
      - nodes
      x-openapi-router-controller: api.controllers.nodes_controller
...
components:
...
  schemas:
...
    Link:
      example:
        BeginNodeID: 0
        Length: 1.4658129805029452
        ID: 0
        EndNodeID: 6
      properties:
        ID:
          description: none
          example: 0
          nullable: false
          readOnly: true
          title: ID
          type: integer
        BeginNodeID:
          description: none
          title: BeginNodeID
          type: integer
        EndNodeID:
          description: none
          title: EndNodeID
          type: integer
        Length:
          description: none
          title: Length
          type: number
      title: Link
      type: object
...
    Node:
      example:
        Type: Type
        EditorParams:
          PosX: 6
          PosY: 1
        ID: 0
        TrendID: 0
        Name: Name
      properties:
        ID:
          description: none
          example: 0
          readOnly: true
          title: ID
          type: integer
        Type:
          description: none
          nullable: false
          pattern: ^.{0,6}
          title: Type
          type: string
        TrendID:
          description: none
          nullable: true
          title: TrendID
          type: integer
        Name:
          description: none
          pattern: ^.{0,50}
          title: Name
          type: string
        EditorParams:
          $ref: '#/components/schemas/EditorNode'
      required:
      - Type
      title: Node
      type: object
...
    EditorNode:
      example:
        PosX: 6
        PosY: 1
      properties:
        PosX:
          description: none
          nullable: false
          pattern: ^.{0,6}
          title: PosX
          type: integer
        PosY:
          description: none
          nullable: false
          title: PosY
          type: integer
      required:
      - PosX
      - PosY
      title: EditorNode
      type: object
    Error:
      properties:
        code:
          format: int32
          title: code
          type: integer
        message:
          title: message
          type: string
      required:
      - code
      - message
      title: Error
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http